tweak copy and rename to /partners
