+++ .github/FUNDING.yml
@@ -1,1 +1,1 @@
-custom: ["https://tailwindcss.com/sponsor"]
+custom: ["https://tailwindcss.com/partners"]
