+++ src/app/partners/adam-teaching.jpg
