+++ src/app/partners/layout.tsx
