+++ src/app/partners/opengraph-image.jpg
