+++ src/app/partners/sponsors.tsx
